api/requirements.txt
astroid==1.6.4
atomicwrites==1.3.0
attrs==19.1.0
awscli==1.16.69
Babel==2.5.3
beautifulsoup4==4.6.0
botocore==1.12.59
certifi==2019.3.9
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.3
cycler==0.10.0
docutils==0.14
falcon==1.4.1
Flask==1.0.2
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==2.1.2
mccabe==0.6.1
more-itertools==7.0.0
msgpack-python==0.5.6
numpy==1.14.0
pluggy==0.9.0
pony==0.7.10
prompt-toolkit==2.0.0
py==1.8.0
pyasn1==0.4.4
pylint==1.9.1
pyparsing==2.2.0
pytest==4.4.0
pytest-cov==2.6.1
python-dateutil==2.6.1
python-mimeparse==1.6.0
pytz==2018.4
PyYAML==3.13
redis==3.0.1
requests==2.21.0
requests-mock==1.5.2
rsa==3.4.2
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
virtualenv==15.1.0
virtualenvwrapper-win==1.2.5
waitress==1.2.1
wcwidth==0.1.7
Werkzeug==0.14.1
wrapt==1.10.11
